完成 CSP1D C3 主问题 MILP 验收收尾
在 application 层完成 C3 主问题 MILP 建模收口，显式注入 ColumnGenerationSolver，并同步 suspend API 调用路径。更新应用验收测试和相关 domain 测试命名，清理 candidatePlans 等遗留门禁命中，补充 daily.md 的验证记录与下一会话交接信息。

+++ ospf-kotlin-framework-csp1d/csp1d-domain-cutting-plan-generation-context/src/test/kotlin/fuookami/ospf/kotlin/framework/csp1d/domain/cutting_plan_generation/service/DFSGeneratorTest.kt
@@ -80,8 +80,8 @@ class DFSGeneratorTest {
 
         assertTrue(plans.isNotEmpty())
         for (plan in plans) {
-            val totalAmount = plan.slices.sumOf { it.amount }
-            assertTrue(totalAmount <= UInt64(2UL), "Total amount should not exceed knife count")
+            val totalAmount = plan.slices.sumOf { it.amount.toInt() }
+            assertTrue(totalAmount <= 2, "Total amount should not exceed knife count")
         }
     }
