+++ ospf-kotlin-framework-csp1d/csp1d-domain-cutting-plan-generation-context/src/test/kotlin/fuookami/ospf/kotlin/framework/csp1d/domain/cutting_plan_generation/service/NSumGeneratorTest.kt
@@ -54,8 +54,8 @@ class NSumGeneratorTest {
 
         assertTrue(plans.isNotEmpty())
         for (plan in plans) {
-            val totalAmount = plan.slices.sumOf { it.amount }
-            assertTrue(totalAmount <= UInt64(2UL), "Total slices should not exceed maxDepth")
+            val totalAmount = plan.slices.sumOf { it.amount.toInt() }
+            assertTrue(totalAmount <= 2, "Total slices should not exceed maxDepth")
         }
     }
